+++ src/scripts/setup/addGitignores.ts
@@ -1,5 +1,4 @@
 import { logStep } from './utils.js';
-import { gitignorePatterns as editorGitignorePatterns } from '../../configs/editors/gitignore.js';
 import { gitignorePatterns as eslintGitignorePatterns } from '../../configs/eslint/gitignore.js';
 import { addGitignorePatterns } from '../utils/gitignoreUtils.js';
 
@@ -12,7 +11,6 @@ const configGitignorePatterns: Record<ConfigType, readonly string[]> = {
     eslint: eslintGitignorePatterns,
     prettier: [], // No patterns for prettier
     typescript: [], // No patterns for typescript
-    cursorSettings: editorGitignorePatterns,
 };
 
 /**